# backend/contest_training_system/docs/api.yml
openapi: 3.0.0
info:
  title: 竞赛集训系统 API
  description: '根据提供的Markdown文档生成的完整版竞赛集训系统RESTful API'
  version: '1.0.0'
servers:
  - url: https://api.example.com
    description: '生产服务器'
...
paths:
...
  /contest/create:
    post:
      tags:
        - '2. 赛事管理'
      summary: '2.1 新建赛事'
      description: '教师新建赛事'
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - startTime
                - end_time
                - visibility
                - problemIds
              properties:
                title:
                  type: string
                  example: '2025年春季赛'
                description:
                  type: string
                  example: '春季编程竞赛'
                startTime:
                  type: string
                  format: date-time
                  example: '2025-04-01 09:00:00'
                end_time:
                  type: string
                  format: date-time
                  example: '2025-04-01 12:00:00'
                password:
                  type: string
                  example: 'abc123'
                visibility:
                  type: string
                  enum: [PUBLIC, PRIVATE]
                  example: 'PUBLIC'
                problemIds:
                  type: array
                  items:
                    type: integer
                  example: [1, 2, 3]
      responses:
        '200':
          description: '创建成功'
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 1
                  msg:
                    type: string
                    example: "创建成功"
                  data:
                    type: object
                    properties:
                      contestId:
                        type: integer
                        example: 101
  /contest/{contestId}:
    get:
      tags:
        - '2. 赛事管理'
      summary: '2.5 获取赛事详情'
      description: '获取赛事详细信息和题目列表'
      security:
        - bearerAuth: []
      parameters:
        - name: contestId
          in: path
          required: true
          schema:
            type: integer
          example: 101
      responses:
        '200':
          description: '获取成功'
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 1
                  msg:
                    type: string
                    example: "success"
                  data:
                    type: object
                    properties:
                      contestId:
                        type: integer
                      title:
                        type: string
                      description:
                        type: string
                      startTime:
                        type: string
                      end_time:
                        type: string
                      visibility:
                        type: string
                      state:
                        type: string
                      creatorId:
                        type: integer
                      problems:
                        type: array
                        items:
                          type: object
                          properties:
                            problemId:
                              type: integer
                            title:
                              type: string
                            displayOrder:
                              type: integer
                            timeLimit:
                              type: integer
                            memoryLimit:
                              type: integer
                      participantCount:
                        type: integer
    delete:
      tags:
        - '2. 赛事管理'
      summary: '2.2 删除赛事'
      description: '根据ID删除赛事'
      security:
        - bearerAuth: []
      parameters:
        - name: contestId
          in: path
          required: true
          schema:
            type: integer
          example: 101
      responses:
        '200':
          description: '删除成功'
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 1
                  msg:
                    type: string
                    example: "删除成功"
                  data:
                    type: 'object'
                    nullable: true
  /contest/update:
    put:
      tags:
        - '2. 赛事管理'
      summary: '2.3 修改赛事'
      description: '教师修改赛事信息'
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - contestId
              properties:
                contestId:
                  type: integer
                  example: 101
                title:
                  type: string
                  example: '2025年春季赛（修改版）'
                description:
                  type: string
                  example: '春季编程竞赛修改版'
                startTime:
                  type: string
                  format: date-time
                end_time:
                  type: string
                  format: date-time
                password:
                  type: string
                visibility:
                  type: string
                  enum: [PUBLIC, PRIVATE]
                problemIds:
                  type: array
                  items:
                    type: integer
                  example: [1,2,3,4]
      responses:
        '200':
          description: '修改成功'
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 1
                  msg:
                    type: string
                    example: "修改成功"
                  data:
                    type: object
                    properties:
                      contestId:
                        type: integer
                        example: 101
                      title:
                        type: string
                        example: '2025年春季赛（修改版）'
  /contest/list:
    get:
      tags:
        - '2. 赛事管理'
      summary: '2.4 查询赛事列表'
      description: '查询赛事列表，支持分页、筛选和排序'
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
        - name: visibility
          in: query
          schema:
            type: string
            enum: [PUBLIC, PRIVATE]
        - name: state
          in: query
          schema:
            type: string
        - name: keyword
          in: query
          schema:
            type: string
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [id_asc, id_desc, title_asc, title_desc, start_time_asc, start_time_desc]
            default: 'id_asc'
      responses:
        '200':
          description: '查询成功'
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 1
                  msg:
                    type: string
                    example: "success"
                  data:
                    type: object
                    properties:
                      total:
                        type: integer
                        example: 50
                      page:
                        type: integer
                        example: 1
                      pageSize:
                        type: integer
                        example: 10
                      contests:
                        type: array
                        items:
                          type: object
                          properties:
                            contestId:
                              type: integer
                            title:
                              type: string
                            description:
                              type: string
                            startTime:
                              type: string
                            end_time:
                              type: string
                            visibility:
                              type: string
                            state:
                              type: string
                            participantCount:
                              type: integer
  /contest/join:
    post:
      tags:
        - '2. 赛事管理'
      summary: '2.6 参与赛事'
      description: '学生参与赛事（公开赛事或输入密码）'
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - contestId
              properties:
                contestId:
                  type: integer
                  example: 101
                password:
                  type: string
                  example: 'abc123'
                teamName:
                  type: string
                  example: '我的队伍'
      responses:
        '200':
          description: '参赛成功'
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 1
                  msg:
                    type: string
                    example: '参赛成功'
                  data:
                    type: 'object'
                    nullable: true
  
  # 3. 题库管理
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: '在需要鉴权的接口的请求头中添加 "Authorization: Bearer <YOUR_TOKEN>"'